installed Cucumber and created login test using Gherkin
