--- src/test/java/MyStepdefs.java
@@ -1,33 +1,0 @@
-import io.cucumber.java.en.And;
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-
-public class MyStepdefs {
-    @Given("I open browser")
-    public void iOpenBrowser() {
-    }
-
-    @And("I open Login page")
-    public void iOpenLoginPage() {
-        
-    }
-
-    @When("I enter email {string}'")
-    public void iEnterEmail(String arg0) {
-        
-    }
-
-    @And("I enter password {string}")
-    public void iEnterPassword(String arg0) {
-    }
-
-    @And("I submit")
-    public void iSubmit() {
-        
-    }
-
-    @Then("I am logged in")
-    public void iAmLoggedIn() {
-    }
-}
