+++ src/test/java/stepDefinition/LoginStepDefinition.java
@@ -1,0 +1,62 @@
+package stepDefinition;
+
+import io.cucumber.java.en.And;
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.testng.Assert;
+
+import java.time.Duration;
+
+public class LoginStepDefinition {
+    WebDriver driver;
+    WebDriverWait wait;
+    @Given("I open browser")
+    public void iOpenBrowser() {
+        WebDriverManager.chromedriver().setup();
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("--disable-notifications");
+        options.addArguments("--remote-allow-origins=*");
+        driver = new ChromeDriver(options);
+        wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+    }
+
+    @And("I open Login page")
+    public void iOpenLoginPage() {
+        driver.get("https:/qa.koel.app/");
+
+    }
+
+    @When("I enter email {string}'")
+    public void iEnterEmail(String email) {
+        wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("[type='email']"))).sendKeys(email);
+
+    }
+
+    @And("I enter password {string}")
+    public void iEnterPassword(String password) {
+        wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("[type='password']"))).sendKeys(password);
+
+    }
+
+    @And("I submit")
+    public void iSubmit() {
+        wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("[type='submit']"))).click();
+
+
+    }
+
+    @Then("I am logged in")
+    public void iAmLoggedIn() {
+        Assert.assertTrue(wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("img.avatar"))).isDisplayed());
+    }
+}
+
+
